requirements.txt
pyfiglet
pycryptodomex
pywin32
sqlite3
tkinter